+++ packages/terra-clinical-site/src/examples/detail-view/DetailViewDividedSmallerTitles.jsx
@@ -24,8 +24,8 @@ const detail2item4 = (<LabelValueView label="Admin Notes" textValue="Try med for
 
 const DetailViewDivided = () => (
   <DetailView
-    title="M. V. I. Adult 10 mL = 10 mL"
-    secondaryTitles={['Sodium bicarbonate 8.4% 50 mL', 'Dextrose 5% in Water 1000 mL']}
+    title="Multiple Ingredients"
+    secondaryTitles={['M. V. I. Adult 10 mL = 10 mL', 'Sodium bicarbonate 8.4% 50 mL', 'Dextrose 5% in Water 1000 mL']}
     subtitles={['30 mg, Oral, Start Date/Time: 06/28/17 20:00:00 CDT. Stop Date/Time: 06/28/17 20:00:00 CDT']}
     accessory={<ItemDisplay text="Ordered" />}
     details={[