+++ packages/terra-clinical-detail-view/CHANGELOG.md
@@ -4,9 +4,11 @@ ChangeLog
 Unreleased
 ----------
 ### Added
-* props "secondaryTitles", "accessory" and "isSmallerTitles" are added.
+* Added new props "secondaryTitles", "accessory", and "isSmallerTitles"
+
 ### Changed
-* Updated spacing & text sizes.
+* Updated spacing & text sizes
+* Updated empty/none default prop values to undefined
 
 2.2.0 - (March 27, 2018)
 ----------